move various methods from common
